Modularize views and models
