+++ FastingApp/Views/Home/HomeView.swift
@@ -1,0 +1,56 @@
+import SwiftUI
+
+struct HomeView: View {
+    @EnvironmentObject var vm: FastingViewModel
+    @State private var now = Date()
+
+    var body: some View {
+        VStack(spacing: 16) {
+            Text(vm.status == .fasting ? "Status: FASTING" : "Status: EATING")
+                .font(.headline)
+                .frame(maxWidth: .infinity, alignment: .leading)
+
+            ProgressRing(progress: vm.progress(now: now))
+                .frame(width: 160, height: 160)
+                .padding(.vertical, 4)
+
+            Group {
+                if vm.status == .fasting, let s = vm.active {
+                    Text("Remaining: \(FastingViewModel.hmsString(from: s.scheduledEnd.timeIntervalSince(now)))")
+                    Text("Start: \(timeString(s.start))  |  End: \(timeString(s.scheduledEnd))")
+                        .font(.footnote)
+                        .foregroundStyle(.secondary)
+                    Button(role: .destructive) { vm.endFast() } label: { Text("Stop Fast") }
+                        .buttonStyle(.borderedProminent)
+                    HStack {
+                        Button("Snooze \(vm.reminders.snoozeMinutes)m") { Task { try await vm.snooze() } }
+                        .buttonStyle(.bordered)
+                    }
+                } else {
+                    Text("Time until fasting: \(vm.remainingString())")
+                    Button { Task { await vm.startFast() } } label: { Text("Start Fast") }
+                        .buttonStyle(.borderedProminent)
+                }
+            }
+            Spacer()
+        }
+        .padding()
+        .onReceive(Timer.publish(every: 1, on: .main, in: .common).autoconnect()) { now = $0 }
+        .navigationTitle("Fasting")
+    }
+
+    func timeString(_ d: Date) -> String {
+        let f = DateFormatter()
+        f.timeStyle = .short
+        f.dateStyle = .none
+        return f.string(from: d)
+    }
+}
+
+#Preview("Home (Fasting)") {
+    let vm = FastingViewModel()
+    let start = Date().addingTimeInterval(-60*60*10)
+    vm.onboarded = true
+    Task { await vm.startFast(now: start) }
+    return HomeView().environmentObject(vm)
+}
